+++ HISTORY.md
@@ -7,11 +7,11 @@
 - [x] define all processor parameters => **all but effect specific parameters mapped**
 - [ ] write get/Set state information for all processor parameters
 - [x] continue definition of EffectComponent & OptionsComponent
-- [ ] define SequenceSequencerComponent
+- [x] define SequenceSequencerComponent **to be continued-methods to be added**
 - [ ] update processor gainpPatterns while modifying patterns on GUI (**crashes**)
 - [ ] cleaner way to pass colortemplates
 #### BUGS
-- [ ] setting the textbutton label color make it crash (why???)
+- [ ] setting the textbutton label color(savePresetButton in options component object) make it crash (why???)
 
 ### v0.01 21092023 Init project
 #### TODO